Add initial PS4 component
This component supports PS4 configuration via YAML. The rest of the work is delegated to the core component. The component also implements the same UI flow by delegating it.

This assumes that the user has a token; which requires registering. As such, the next step would be adding an offline flow to get the token and store it in a file or show in the screen. Although, at the moment, the methods can be combined.